add some unit test for delete workflowTask

+++ QuickServiceWebAPI/Services/Implements/WorkflowTaskService.cs
@@ -125,7 +125,10 @@ namespace QuickServiceWebAPI.Services.Implements
                 throw new AppException($"Workflow Task with id {workflowTaskId} not found");
             }
             var listWorkflowAssignment = await _workflowAssignmentRepository.GetWorkflowAssignmentsByWorkflowTaskId(workflowTaskId);
-            await _workflowAssignmentService.DeleteListWorkflowAssignment(listWorkflowAssignment);
+            if (listWorkflowAssignment.IsAny())
+            {
+                await _workflowAssignmentService.DeleteListWorkflowAssignment(listWorkflowAssignment);
+            }
             await _repository.DeleteWorkflowTask(workflowTask);
         }
